fix: make sample size an integer in fmt command
Previously, sample size was deserialized as an f64, leading to values
like "15000.0", which cause failures when running other commands like
`sumher_rs gcov`, where an integer is expected and the period character
is unexpected.

+++ src/util.rs
@@ -54,7 +54,7 @@ struct PlinkSumstats {
     #[serde(rename(serialize = "A2", deserialize = "OMITTED"))]
     a2: String,
     #[serde(rename(serialize = "n", deserialize = "OBS_CT"))]
-    n: f64,
+    n: i64,
     #[serde(rename(serialize = "Z", deserialize = "T_STAT"))]
     z: f64,
 }
